bpf: Move host firewall code to header file
[ upstream commit ffa49f710ccd4c7bf2dd373f95c8eeb31f31d813 ]

The host firewall functions will be called from nodeport code in next
commit. A few minor changes were also made in the code moved to please
checkpatch.
