+++ bpf/bpf_host.c
@@ -45,10 +45,7 @@
 #include "lib/lb.h"
 #include "lib/nodeport.h"
 #include "lib/eps.h"
-#ifdef ENABLE_HOST_FIREWALL
-# include "lib/policy.h"
-# include "lib/policy_log.h"
-#endif
+#include "lib/host_firewall.h"
 
 #if defined(ENABLE_IPV4) || defined(ENABLE_IPV6)
 static __always_inline int rewrite_dmac_to_host(struct __ctx_buff *ctx,
@@ -156,173 +153,6 @@ resolve_srcid_ipv6(struct __ctx_buff *ctx, __u32 srcid_from_proxy,
 
 	return src_id;
 }
-
-#ifdef ENABLE_HOST_FIREWALL
-static __always_inline int
-ipv6_host_policy_egress(struct __ctx_buff *ctx, __u32 srcID)
-{
-	int ret, verdict, l3_off = ETH_HLEN, l4_off, hdrlen;
-	struct ct_state ct_state_new = {}, ct_state = {};
-	__u8 policy_match_type = POLICY_MATCH_NONE;
-	__u8 audited = 0;
-	struct remote_endpoint_info *info;
-	struct ipv6_ct_tuple tuple = {};
-	__u32 dstID = 0, monitor = 0;
-	union v6addr orig_dip;
-	void *data, *data_end;
-	struct ipv6hdr *ip6;
-
-	/* Only enforce host policies for packets from host IPs. */
-	if (srcID != HOST_ID)
-		return CTX_ACT_OK;
-
-	if (!revalidate_data(ctx, &data, &data_end, &ip6))
-		return DROP_INVALID;
-
-	/* Lookup connection in conntrack map. */
-	tuple.nexthdr = ip6->nexthdr;
-	ipv6_addr_copy(&tuple.daddr, (union v6addr *) &ip6->daddr);
-	ipv6_addr_copy(&orig_dip, (union v6addr *) &ip6->daddr);
-	hdrlen = ipv6_hdrlen(ctx, ETH_HLEN, &tuple.nexthdr);
-	if (hdrlen < 0)
-		return hdrlen;
-	l4_off = l3_off + hdrlen;
-	ret = ct_lookup6(get_ct_map6(&tuple), &tuple, ctx, l4_off, CT_EGRESS,
-			 &ct_state, &monitor);
-	if (ret < 0)
-		return ret;
-
-	/* Retrieve destination identity. */
-	info = lookup_ip6_remote_endpoint(&orig_dip);
-	if (info != NULL && info->sec_label)
-		dstID = info->sec_label;
-	cilium_dbg(ctx, info ? DBG_IP_ID_MAP_SUCCEED6 : DBG_IP_ID_MAP_FAILED6,
-		   orig_dip.p4, dstID);
-
-	/* Perform policy lookup. */
-	verdict = policy_can_egress6(ctx, &tuple, srcID, dstID,
-				     &policy_match_type, &audited);
-
-	/* Reply traffic and related are allowed regardless of policy verdict. */
-	if (ret != CT_REPLY && ret != CT_RELATED && verdict < 0) {
-		send_policy_verdict_notify(ctx, dstID, tuple.dport,
-					   tuple.nexthdr, POLICY_EGRESS, 1,
-					   verdict, policy_match_type, audited);
-		return verdict;
-	}
-
-	switch (ret) {
-	case CT_NEW:
-		send_policy_verdict_notify(ctx, dstID, tuple.dport,
-					   tuple.nexthdr, POLICY_EGRESS, 1,
-					   verdict, policy_match_type, audited);
-		ct_state_new.src_sec_id = HOST_ID;
-		ret = ct_create6(get_ct_map6(&tuple), &CT_MAP_ANY6, &tuple,
-				 ctx, CT_EGRESS, &ct_state_new, verdict > 0);
-		if (IS_ERR(ret))
-			return ret;
-		break;
-
-	case CT_ESTABLISHED:
-	case CT_RELATED:
-	case CT_REPLY:
-		break;
-
-	default:
-		return DROP_UNKNOWN_CT;
-	}
-
-	return CTX_ACT_OK;
-}
-
-static __always_inline int
-ipv6_host_policy_ingress(struct __ctx_buff *ctx, __u32 *srcID)
-{
-	struct ct_state ct_state_new = {}, ct_state = {};
-	__u8 policy_match_type = POLICY_MATCH_NONE;
-	__u8 audited = 0;
-	__u32 monitor = 0, dstID = WORLD_ID;
-	struct remote_endpoint_info *info;
-	int ret, verdict, l4_off, hdrlen;
-	struct ipv6_ct_tuple tuple = {};
-	union v6addr orig_sip;
-	void *data, *data_end;
-	struct ipv6hdr *ip6;
-
-	if (!revalidate_data(ctx, &data, &data_end, &ip6))
-		return DROP_INVALID;
-
-	/* Retrieve destination identity. */
-	ipv6_addr_copy(&tuple.daddr, (union v6addr *) &ip6->daddr);
-	info = lookup_ip6_remote_endpoint(&tuple.daddr);
-	if (info != NULL && info->sec_label)
-		dstID = info->sec_label;
-	cilium_dbg(ctx, info ? DBG_IP_ID_MAP_SUCCEED6 : DBG_IP_ID_MAP_FAILED6,
-		   tuple.daddr.p4, dstID);
-
-	/* Only enforce host policies for packets to host IPs. */
-	if (dstID != HOST_ID)
-		return CTX_ACT_OK;
-
-	/* Lookup connection in conntrack map. */
-	tuple.nexthdr = ip6->nexthdr;
-	ipv6_addr_copy(&tuple.saddr, (union v6addr *) &ip6->saddr);
-	ipv6_addr_copy(&orig_sip, (union v6addr *) &ip6->saddr);
-	hdrlen = ipv6_hdrlen(ctx, ETH_HLEN, &tuple.nexthdr);
-	if (hdrlen < 0)
-		return hdrlen;
-	l4_off = ETH_HLEN + hdrlen;
-	ret = ct_lookup6(get_ct_map6(&tuple), &tuple, ctx, l4_off, CT_INGRESS,
-			 &ct_state, &monitor);
-	if (ret < 0)
-		return ret;
-
-	/* Retrieve source identity. */
-	info = lookup_ip6_remote_endpoint(&orig_sip);
-	if (info != NULL && info->sec_label)
-		*srcID = info->sec_label;
-	cilium_dbg(ctx, info ? DBG_IP_ID_MAP_SUCCEED6 : DBG_IP_ID_MAP_FAILED6,
-		   orig_sip.p4, *srcID);
-
-	/* Perform policy lookup */
-	verdict = policy_can_access_ingress(ctx, *srcID, dstID, tuple.dport,
-					    tuple.nexthdr, false,
-					    &policy_match_type, &audited);
-
-	/* Reply traffic and related are allowed regardless of policy verdict. */
-	if (ret != CT_REPLY && ret != CT_RELATED && verdict < 0) {
-		send_policy_verdict_notify(ctx, *srcID, tuple.dport,
-					   tuple.nexthdr, POLICY_INGRESS, 1,
-					   verdict, policy_match_type, audited);
-		return verdict;
-	}
-
-	switch (ret) {
-	case CT_NEW:
-		send_policy_verdict_notify(ctx, *srcID, tuple.dport,
-					   tuple.nexthdr, POLICY_INGRESS, 1,
-					   verdict, policy_match_type, audited);
-
-		/* Create new entry for connection in conntrack map. */
-		ct_state_new.src_sec_id = *srcID;
-		ct_state_new.node_port = ct_state.node_port;
-		ret = ct_create6(get_ct_map6(&tuple), &CT_MAP_ANY6, &tuple,
-				 ctx, CT_INGRESS, &ct_state_new, verdict > 0);
-		if (IS_ERR(ret))
-			return ret;
-
-	case CT_ESTABLISHED:
-	case CT_RELATED:
-	case CT_REPLY:
-		break;
-
-	default:
-		return DROP_UNKNOWN_CT;
-	}
-
-	return CTX_ACT_OK;
-}
-#endif /* ENABLE_HOST_FIREWALL */
 
 static __always_inline int
 handle_ipv6(struct __ctx_buff *ctx, __u32 secctx, const bool from_host)
@@ -601,172 +431,6 @@ resolve_srcid_ipv4(struct __ctx_buff *ctx, __u32 srcid_from_proxy,
 
 	return src_id;
 }
-
-#ifdef ENABLE_HOST_FIREWALL
-static __always_inline int
-ipv4_host_policy_egress(struct __ctx_buff *ctx, __u32 srcID)
-{
-	struct ct_state ct_state_new = {}, ct_state = {};
-	int ret, verdict, l4_off, l3_off = ETH_HLEN;
-	__u8 policy_match_type = POLICY_MATCH_NONE;
-	__u8 audited = 0;
-	struct remote_endpoint_info *info;
-	struct ipv4_ct_tuple tuple = {};
-	__u32 dstID = 0, monitor = 0;
-	void *data, *data_end;
-	struct iphdr *ip4;
-
-	/* Only enforce host policies for packets from host IPs. */
-	if (srcID != HOST_ID)
-		return CTX_ACT_OK;
-
-	if (!revalidate_data(ctx, &data, &data_end, &ip4))
-		return DROP_INVALID;
-
-	/* Lookup connection in conntrack map. */
-	tuple.nexthdr = ip4->protocol;
-	tuple.daddr = ip4->daddr;
-	tuple.saddr = ip4->saddr;
-	l4_off = l3_off + ipv4_hdrlen(ip4);
-	ret = ct_lookup4(get_ct_map4(&tuple), &tuple, ctx, l4_off, CT_EGRESS,
-			 &ct_state, &monitor);
-	if (ret < 0)
-		return ret;
-
-	/* Retrieve destination identity. */
-	info = lookup_ip4_remote_endpoint(ip4->daddr);
-	if (info != NULL && info->sec_label)
-		dstID = info->sec_label;
-	cilium_dbg(ctx, info ? DBG_IP_ID_MAP_SUCCEED4 : DBG_IP_ID_MAP_FAILED4,
-		   ip4->daddr, dstID);
-
-	/* Perform policy lookup. */
-	verdict = policy_can_egress4(ctx, &tuple, srcID, dstID,
-				     &policy_match_type, &audited);
-
-	/* Reply traffic and related are allowed regardless of policy verdict. */
-	if (ret != CT_REPLY && ret != CT_RELATED && verdict < 0) {
-		send_policy_verdict_notify(ctx, dstID, tuple.dport,
-					   tuple.nexthdr, POLICY_EGRESS, 0,
-					   verdict, policy_match_type, audited);
-		return verdict;
-	}
-
-	switch (ret) {
-	case CT_NEW:
-		send_policy_verdict_notify(ctx, dstID, tuple.dport,
-					   tuple.nexthdr, POLICY_EGRESS, 0,
-					   verdict, policy_match_type, audited);
-		ct_state_new.src_sec_id = HOST_ID;
-		ret = ct_create4(get_ct_map4(&tuple), &CT_MAP_ANY4, &tuple,
-				 ctx, CT_EGRESS, &ct_state_new, verdict > 0);
-		if (IS_ERR(ret))
-			return ret;
-		break;
-
-	case CT_ESTABLISHED:
-	case CT_RELATED:
-	case CT_REPLY:
-		break;
-
-	default:
-		return DROP_UNKNOWN_CT;
-	}
-
-	return CTX_ACT_OK;
-}
-
-static __always_inline int
-ipv4_host_policy_ingress(struct __ctx_buff *ctx, __u32 *srcID)
-{
-	struct ct_state ct_state_new = {}, ct_state = {};
-	int ret, verdict, l4_off, l3_off = ETH_HLEN;
-	__u8 policy_match_type = POLICY_MATCH_NONE;
-	__u8 audited = 0;
-	__u32 monitor = 0, dstID = WORLD_ID;
-	struct remote_endpoint_info *info;
-	struct ipv4_ct_tuple tuple = {};
-	bool is_untracked_fragment = false;
-	void *data, *data_end;
-	struct iphdr *ip4;
-
-	if (!revalidate_data(ctx, &data, &data_end, &ip4))
-		return DROP_INVALID;
-
-	/* Retrieve destination identity. */
-	info = lookup_ip4_remote_endpoint(ip4->daddr);
-	if (info != NULL && info->sec_label)
-		dstID = info->sec_label;
-	cilium_dbg(ctx, info ? DBG_IP_ID_MAP_SUCCEED4 : DBG_IP_ID_MAP_FAILED4,
-		   ip4->daddr, dstID);
-
-	/* Only enforce host policies for packets to host IPs. */
-	if (dstID != HOST_ID)
-		return CTX_ACT_OK;
-
-	/* Lookup connection in conntrack map. */
-	tuple.nexthdr = ip4->protocol;
-	tuple.daddr = ip4->daddr;
-	tuple.saddr = ip4->saddr;
-	l4_off = l3_off + ipv4_hdrlen(ip4);
-#ifndef IPV4_FRAGMENTS
-	/* Indicate that this is a datagram fragment for which we cannot
-	 * retrieve L4 ports. Do not set flag if we support fragmentation.
-	 */
-	is_untracked_fragment = ipv4_is_fragment(ip4);
-#endif
-	ret = ct_lookup4(get_ct_map4(&tuple), &tuple, ctx, l4_off, CT_INGRESS,
-			 &ct_state, &monitor);
-	if (ret < 0)
-		return ret;
-
-	/* Retrieve source identity. */
-	info = lookup_ip4_remote_endpoint(ip4->saddr);
-	if (info != NULL && info->sec_label)
-		*srcID = info->sec_label;
-	cilium_dbg(ctx, info ? DBG_IP_ID_MAP_SUCCEED4 : DBG_IP_ID_MAP_FAILED4,
-		   ip4->saddr, *srcID);
-
-	/* Perform policy lookup */
-	verdict = policy_can_access_ingress(ctx, *srcID, dstID, tuple.dport,
-					    tuple.nexthdr,
-					    is_untracked_fragment,
-					    &policy_match_type, &audited);
-
-	/* Reply traffic and related are allowed regardless of policy verdict. */
-	if (ret != CT_REPLY && ret != CT_RELATED && verdict < 0) {
-		send_policy_verdict_notify(ctx, *srcID, tuple.dport,
-					   tuple.nexthdr, POLICY_INGRESS, 0,
-					   verdict, policy_match_type, audited);
-		return verdict;
-	}
-
-	switch (ret) {
-	case CT_NEW:
-		send_policy_verdict_notify(ctx, *srcID, tuple.dport,
-					   tuple.nexthdr, POLICY_INGRESS, 0,
-					   verdict, policy_match_type, audited);
-
-		/* Create new entry for connection in conntrack map. */
-		ct_state_new.src_sec_id = *srcID;
-		ct_state_new.node_port = ct_state.node_port;
-		ret = ct_create4(get_ct_map4(&tuple), &CT_MAP_ANY4, &tuple,
-				 ctx, CT_INGRESS, &ct_state_new, verdict > 0);
-		if (IS_ERR(ret))
-			return ret;
-
-	case CT_ESTABLISHED:
-	case CT_RELATED:
-	case CT_REPLY:
-		break;
-
-	default:
-		return DROP_UNKNOWN_CT;
-	}
-
-	return CTX_ACT_OK;
-}
-#endif /* ENABLE_HOST_FIREWALL */
 
 static __always_inline int
 handle_ipv4(struct __ctx_buff *ctx, __u32 secctx, const bool from_host)
